pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["lekhins <alekseenkolesha@yandex.ru>"]
readme = "README.md"
packages = [{include = 'page_analyzer'}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
requests = "^2.32.3"
load-dotenv = "^0.1.0"
bs4 = "^0.0.2"
psycopg2 = "^2.9.9"
jinja2 = "^3.1.4"
validators = "^0.28.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
